fixed bugs in code result of merge conflict

+++ app/src/main/java/com/group21/sneakerhub/model/Adidas.java
@@ -2,7 +2,7 @@ package com.group21.sneakerhub.model;
 
 public class Adidas extends Category {
 
-    public Adidas(String name, int Id, String URI, String colour, String layout) {
+    public Adidas(String name, long Id, String URI, String colour, String layout) {
         super(name, Id, URI, colour,layout);
     }
