+++ app/src/main/java/com/group21/sneakerhub/model/Entity.java
@@ -3,13 +3,13 @@ package com.group21.sneakerhub.model;
 public abstract class Entity {
 
     public String name;
-    public int id;
+    public long id;
 
     public String GetName() {
         return this.name;
     }
 
-    public int GetId(){
+    public long GetId(){
         return this.id;
     }
 }
