+++ app/src/main/java/com/group21/sneakerhub/model/Vans.java
@@ -2,7 +2,7 @@ package com.group21.sneakerhub.model;
 
 public class Vans extends Category{
 
-    public Vans(String name, int Id, String URI, String colour, String layout) {
+    public Vans(String name, long Id, String URI, String colour, String layout) {
         super(name, Id, URI, colour, layout);
     }
